moved registration in menu
