+++ app/controllers/api/v1/users_controller.rb
@@ -6,9 +6,9 @@ class API::V1::UsersController < ApplicationController
     @user = User.new(user_params)
 
     if @user.save
-      render :show, status: :created
+      render json: @user, status: :created
     else
-      render_validation_errors(@user)
+      render json: @user.errors, status: :unprocessable_entity
     end
   end
